+++ megamek/resources/megamek/client/messages.properties
@@ -1455,8 +1455,6 @@ CommonSettingsDialog.TraceOverlayScale.tooltip=0-1.5
 CommonSettingsDialog.TraceOverlayOrigin=Board Editor - Trace Overlay - Origin
 CommonSettingsDialog.TraceOverlayOrigin.tooltip=Used to center the image
 CommonSettingsDialog.TraceOverlayImageFile=Board Editor - Trace Overlay - Image File
-CommonSettingsDialog.ToastDripSeconds=Toast Notification - Drip Cadence (seconds)
-CommonSettingsDialog.ToastDripSeconds.tooltip=Seconds between consecutive toast notifications when a single report (e.g., a building collapse with multiple damage events) is split into per-event toasts. Lower = faster scroll, less time to read each. Range 1-10, default 2.
 # Pipe-separated substring list. Any report entry whose stripped preview contains one of these is filtered out of the
 # scrolling toast feed (it remains visible in the report panel). Used to suppress routine round-summary lines that
 # would otherwise dominate the kill-feed during initiative reports. Translators: override per-locale to match the
